requirements.txt
openai
jsonlines
python-dotenv